+++ app/utils/colorMapping.ts
